✅ Unit tests, bug fixes, and inference implementation
Add logic variable parameters to background.py
  (use_std_logic_variables, use_prolog_variables)
Add rdn.RDN.predict and rdn.RDN.predict_proba implementations
Add unit tests for these new features

The big addition is implementations to access the predictions of
BoostSRL through predict() and predict_proba(). Both rely on a
private method for running inference, but differ on the return
value. The former returns an ndarray of 0.0/1.0 class membership
values, whereas the latter returns the probability of belonging
to the positive class.

Notes
-----

The private method for running inference is not well-tested,
and relies on np.readtxt for reading/parsing the
{target}_results.db files.