✅ Update tests after modifying `load_*` interfaces
Instead of returning a scikit-learn Bunch object, this modifies the
interface of `load_toy_cancer` and `load_toy_father` to behave like
`relational_datasets.load`, which returns a `train` and `test` tuple
of a Named Tuple.

Update tests to reflect this change.

Add `compat_tests` directory to track tests for compatibility with
third-party packages.